Add the gzipped file to the context's distFiles when keep is true
By not adding the gzipped files to the distFiles we end up leaving
them out of several things e.g. they're never added to the manifest
for a build and they're never uploaded with the S3 plugin. In much
the same way as the json-config plugin adds the built json file to
the context, this PR will mean the gzip plugin now adds all of the
gzipped files to the context.