Fix invalid atime being set
It's not super clear from the node docs, but it appears that the value
of atime needs to be in seconds since the eppoc, but Date.now returns milliseconds
which sets the timestamp far into the future. This causes node tar to
die when it attempts to encode this time and finds a very large integer.

We can also pass a Date object directly here so that's what I've done.